- normalize helper functions in main binary - return errors, main function handles logging + fatal exit - move aws related code to aws.go for main binary - move datalake related code to datalake.go for main binary
